rename UPS component to UPS screen
