+++ libs/ui/src/diagnostics/uninterruptible_power_supply_screen.tsx
